+++ spec/bigcommerce_oauth_api/client/client_nested_module_spec.rb
@@ -5,15 +5,17 @@ require 'bigcommerce-oauth-api/client'
 
 describe BigcommerceOAuthAPI::Client do
   [
-      { api_module: 'product', api_parent_module: 'order', methods: [:all, :select]},
+      { api_module: 'product', api_parent_module: 'order', methods: [:all, :select, :count]},
       { api_module: 'shipping_address', api_parent_module: 'order', methods: [:all, :select]},
+      { api_module: 'coupon', api_parent_module: 'order', methods: [:all, :select]},
       { api_module: 'message', api_parent_module: 'order', methods: [:all, :select]},
       { api_module: 'shipment', api_parent_module: 'order', methods: [:all, :select, :create, :update, :delete]},
+      { api_module: 'tax', api_parent_module: 'order', methods: [:all, :select]},
       { api_module: 'configurable_field', api_parent_module: 'product', methods: [:all, :select, :delete]},
       { api_module: 'custom_field', api_parent_module: 'product', methods: [:all, :select, :create, :update, :delete]},
       { api_module: 'option', api_parent_module: 'option_set', methods: [:all, :select, :create, :update, :delete]},
       { api_module: 'value', api_parent_module: 'option_set', methods: [:all, :select, :create, :update, :delete]},
-      { api_module: 'discount_rule', api_parent_module: 'product', methods: [:all, :select, :create, :update, :delete]},
+      { api_module: 'discount_rule', api_parent_module: 'product', methods: [:all, :select, :create, :update, :delete, :count]},
       { api_module: 'image', api_parent_module: 'product', methods: [:all, :select, :create, :update, :delete]},
       { api_module: 'option', api_parent_module: 'product', methods: [:all, :select]},
       { api_module: 'review', api_parent_module: 'product', methods: [:all]},