+++ CHANGELOG.md
@@ -4,10 +4,13 @@ Features:
 
     - add legacy api support
     - add omniref and revert to using rdoc
+    - add order coupon apis
+    - add order tax apis
 
 Bugfixes:
 
     - update geography apis
+    - add a few missing "count" apis
 
 ### 1.1.5 (2014-12-25)